PONG-11 Cast result of time call to suppress compiler warning

+++ Pong/Pong/Pong/pong-GuesserPaddleAiDefn.cpp
@@ -5,7 +5,7 @@
 namespace pong {
 
 	GuesserPaddleAi::GuesserPaddleAi() {
-		this->generator.seed(time(NULL));
+		this->generator.seed((unsigned int)time(NULL));
 	}
 
 	PaddleInputType GuesserPaddleAi::resolvePaddleInputType(PaddleAiInput input) {
